Added Delete Button to dialog.xml; Edit Method deleteEntry.

+++ app/src/main/java/com/mkg_dhbw/xcalc/repositories/SQLiteRepository.java
@@ -75,7 +75,7 @@ public class SQLiteRepository extends SQLiteOpenHelper {
     }
 
     // delete Entry
-    public void deleteEntry(int id) {
+    public void deleteEntry(long id) {
         SQLiteDatabase db = this.getWritableDatabase();
         String selection = HistoryRequestEntry._ID + " = ?";
         String[] selectionArgs = { ""+id };
